Internal #861: Refactor Window Boundaries
Move the WindowBoundariesState classes to separate files.
